==> notebooks/0_general_data.ipynb <==
import pandas as pd
import requests
import re

df_oscars = pd.read_csv('../data/raw/the_oscar_award.csv')
df_oscars.head(2)

df_golden = pd.read_csv('../data/raw/golden_globe_awards.csv')
df_golden.head(2)

df_imdb = pd.read_csv('../data/raw/imdb_movie_data_2023.csv')
df_imdb.head(2)

#Nos enfocamos en cada DF por separado para la limpieza primero
#queremos comparar primero los datos de los oscars con los golden globes, que no empezaron hasta 1944, por lo que la información de años anteriores no nos vale
#las categorías de premios de los oscars, además, estaban divididas en 'black and white' y 'color' hasta finales de los años 60, por lo que eliminaremos las filas anteriores a esa fecha (tampoco son relevantes para el estudio)
#los DFs de golden globes y oscars incluyen información de nominados y ganadores. Mantendremos las dos para comprobar si los nominados suelen ser los mismos
#el número de la edición de la ceremonia (columna 'Ceremony') tanto en golden globes como en oscars no nos hace falta teniedo ya el año de la ceremonia
#df_oscars y df_golden comparten una estructura similar de columnas, así que el trabajo de limpieza se centrará en unificar formato y categorías de premios, comprobar que los nombres de películas y actores están escritos igual, etc.
#propuesta --- hacer un count de premios incluyendo todas las categorías, y dejar las categorías específicas en comentarios. Igual sacar columnas para Best movie -- True y Best director -- True como más relevantes para la valoración general de la película
# en el DF de imdb nos interesarán todas las columnas menos PG Rating y duración, que se eliminarán. Del resto, habrá que unificar formato con las columnas de movie name y director y las respectivas de los otros dos df
#si no obtenemos suficientes datos del df de imdb, completaremos con datos de tmdb (api disponible)
