==> notebooks/1_oscars_data_cleaning.ipynb <==
import pandas as pd
import requests
import re
import cleaning_functions as cf

df_oscars = pd.read_csv('../data/raw/the_oscar_award.csv')
df_oscars.head(2)

df_oscars.info()

df_oscars.drop('ceremony',axis=1,inplace=True)

df_oscars.category = df_oscars.category.str.capitalize()

#quitar premios previos a 1967, ya que en este año se realizaron grandes cambios en la estructura de los premios (eliminación de categorías black and white)
df_oscars = df_oscars[df_oscars.year_film >= 1967]
df_oscars

df_oscars.isnull().any()

nan_rows_oscars = df_oscars[df_oscars.film.isnull()]

nan_rows_oscars.info()

#comprobar contenido de las filas con nan values en películas. parece que es por premios especiales y premios honorarios, por lo que las eliminamos después
nan_rows_oscars.tail(5)

df_oscars.dropna(subset='film', inplace = True)

df_oscars.isnull().any()

subset = df_oscars[(df_oscars.year_film == 1967) & (df_oscars.winner == True)]

category_value_drop = ['Award of commendation']

df_oscars = cf.drop_rows_by_value(df_oscars,df_oscars.category,category_value_drop)

changes_category_values = {
    'Actress':'Actress in a leading role',
    'Actor': 'Actor in a leading role',
    'Art direction': 'Production design',
    'Best motion picture': 'Picture',
    'Directing':'Director',
    'Makeup': 'Makeup and hairstyling',
    'Sound':'Sound effects',
    'Sound effects editing':'Sound effects',
    'Sound editing': 'Sound effects',
    'Sound mixing':'Sound effects',
    'Foreign language film': 'International feature film'
}

changes_category_strings = {
    'subject': 'film',
    'cartoon': 'animated',
    '(':'',
    ')':'',
    'Special achievement award ':'',
    'Special ':'',  
    'dramatic ':''
}

df_oscars.category = cf.clean_strings_column(df_oscars.category, changes_category_strings)

df_oscars.category = df_oscars.category.str.capitalize()

df_oscars = cf.clean_values_column(df_oscars,changes_category_values)

#Crear sets con categorías similares para ver qué más se puede limpiar:

music_categories = cf.values_iterable_regex(df_oscars.category,r'\b\w*music\w*\b')
music_categories

music_patterns = {
    r'\b\w*scor\w*\b':'Original score'
}

music_clean = cf.value_replace_dictio_regex(music_categories, music_patterns, 'Original song')

df_oscars = cf.clean_values_column(df_oscars,music_clean)

writing_categories = cf.values_iterable_regex(df_oscars.category,r'\b\w*writing\w*\b')
writing_categories

writting_patterns = {
    r'\b\w*not preciously\w*\b':'Original screenplay',
    r'\b\w*directly\w*\b': 'Original screenplay',
    r'\b\w*original\w*\b': 'Original screenplay'
}

writing_clean = cf.value_replace_dictio_regex(writing_categories, writting_patterns, 'Adapted screenplay')

df_oscars = cf.clean_values_column(df_oscars,writing_clean)

df_oscars.nunique()

df_oscars.category.unique()

df_oscars.reset_index(drop = True, inplace = True)

df_oscars.head(2)

subset_b = df_oscars[(df_oscars.year_film == 1995) & (df_oscars.category == 'Best picture')]
subset_b.head(60)

#para no perder datos al combinar este dataframe con el de golden globes, cambiabos el nombre de algunas columnas:
df_oscars.rename(columns={"category": "Oscars nominations", 'winner':'Oscars winner'}, inplace=True)
df_oscars.head(2)

df_oscars.to_csv('oscars_cleaned.csv', index = False)
