==> notebooks/2_golden_globes_data_cleaning.ipynb <==
import pandas as pd
import requests
import re
import cleaning_functions as cf

df_golden = pd.read_csv('../data/raw/golden_globe_awards.csv')

df_golden.drop('ceremony',axis=1,inplace=True)
df_golden.head(2)

df_golden.info()

df_golden.category = df_golden.category.str.capitalize()

#eliminamos los mismos años que en oscars
df_golden = df_golden[df_golden.year_film >= 1967]
df_golden

#Parece que hay categorías que no tienen una película vinculada, así que vamos a hacer una primera limpieza de los valores en Category para quitarnos las innecesarias

df_golden.category.unique()

#primero eliminamos las filas con categorías sobre televisión, ya que nuestro estudio se centra en cine/películas
tv_categories = cf.values_iterable_regex(df_golden.category,r'\b\w*television\w*\b')

df_golden = cf.drop_rows_by_value(df_golden,df_golden.category,tv_categories)

#eliminamos también categrías honoríficas no vinculadas a una película concreta, y las categrías "New star" que se eliminaron en los 80:
honorary_awards =['Cecil b. demille award','Henrietta award (world film favorites)','Special achievement award','Carol burnett award','New star of the year - actress', 'New star of the year - actor', 'New star of the year']

df_golden = cf.drop_rows_by_value(df_golden,df_golden.category,honorary_awards)

df_golden.reset_index(drop = True)

df_golden.isnull().sum()

nan_rows_golden = df_golden[df_golden.film.isnull()]

#parece que el nombre de las películas nominadas a premios de películas aparecen en nominee y no en film
nan_rows_golden

df_golden.film.fillna(df_golden.nominee, inplace=True)

df_golden.isnull().sum()

#seguimos limpiando los valores únicos
df_golden.nunique()

df_golden.category.unique()

cat_changes = {
    'Actress in a leading role - musical or comedy': 'Best actress - musical or comedy',
    'Best performance by an actress in a motion picture - drama': 'Best actress - drama',
    'Best performance by an actress in a motion picture - musical or comedy': 'Best actress - musical or comedy',
    'Best performance by an actress in a supporting role in any motion picture': 'Best supporting actress - motion picture',
    'Best performance by an actor in a motion picture - drama': 'Best actor - drama',
    'Best performance by an actor in a motion picture - musical or comedy': 'Best actor - musical or comedy',
    'Best performance by an actor in a supporting role in any motion picture': 'Best supporting actor - motion picture',
    'Foreign film - foreign language': 'Best motion picture - foreign language',
    'Foreign film - english language': 'Best motion picture - foreign language'
}

df_golden = cf.clean_values_column (df_golden, cat_changes)
df_golden.category.unique()

df_golden.head(5)

#para no perder datos al combinar este dataframe con el de oscars, cambiabos el nombre de algunas columnas:
df_golden.rename(columns={"category": "Golden Globes nominations", 'win':'Golden Globes winner'}, inplace=True)
df_golden.tail()

#la tabla solo contiene datos hasta 2019/2020, por lo que tendremos que sacar los restantes haciendo scrapping de la página de los Golden Globes

#df_golden.to_csv('golden_cleaned.csv', index = False)
