==> notebooks/3_golden_globes_scrap.ipynb <==
import pandas as pd
import requests
from bs4 import BeautifulSoup
import re

headers = ['year_film','year_award','Golden Globes nominations','nominee','film','Golden Globes winner']

years = [2021,2022,2023]

url = f'https://goldenglobes.com/winners-nominees/?y=2021' #podemos cambiar el año por cada uno de los años en un bucle

for y in years:
    url = f'https://goldenglobes.com/winners-nominees/?y={y}'
    response_test = requests.get(url)
response_test

#comprobamos que el bucle funciona
soup_test = BeautifulSoup(response_test.content, "html.parser")
soup_test.find('div',class_='c-nomination-card__content')

#sacamos cada uno de los elementos que necesitamos individualmente
response = requests.get(url)
response

sopa = BeautifulSoup(response.content, "html.parser")

nom_cards = sopa.find_all('div',class_='c-nomination-card__content') #cada uno de los apartados de premios
#los ganadores tienen una categoría distinta que podemos usar para sacar si han ganado o no
winners = nom_cards[0].find_all('li', class_='c-nomination-card__nominee js-nomination-nominee is-active')[0].find_all('div')[1].find_all('a')
winners[0].text.strip()

nom_cards[0].find_all('li', class_='c-nomination-card__nominee js-nomination-nominee is-active')[0].find_all('div')[1].find_all('a')[1].text.strip()

#sacamos la categoría del premio
category = nom_cards[0].find('h3', class_='u-type-h4 u-mb-2 u-mb-4@md').text.strip()

#sacamos el resto de nominados
nominees_f = nom_cards[0].find_all('li', class_='c-nomination-card__nominee js-nomination-nominee')

nominees_f[0]

nominee = nominees_f[0].find('a', class_='c-nomination-card__name c-nomination-card__link u-color-black-01').text.strip()
nominee

film = nominees_f[0].find('div', class_='u-type-h11').text.strip()
film

#creamos el diccionario del que saldrá el DF y lo completamos con los datos de la web por cada año

golden_globes = {
    'year_film': [],
    'year_award': [],
    'Golden Globes nominations':[],
    'nominee':[],
    'film':[],
    'Golden Globes winner':[]    
}

years = [2021,2022,2023]

for y in years:
    url = f'https://goldenglobes.com/winners-nominees/?y={y}'
    response = requests.get(url)
    sopa = BeautifulSoup(response.content, "html.parser")
    nom_cards = sopa.find_all('div',class_='c-nomination-card__content')
    year_award = y
    year_film = y -1
    for n in range(len(nom_cards)):
        category = nom_cards[n].find('h3', class_='u-type-h4 u-mb-2 u-mb-4@md').text.strip()
        winner_card = nom_cards[n].find_all('li', class_='c-nomination-card__nominee js-nomination-nominee is-active')[0].find_all('div')[1].find_all('a')
        nominee = winner_card[0].text.strip()
        film = ''
        if len(winner_card) > 1:
            film = winner_card[1].text.strip()
        winner = True
        golden_globes['Golden Globes nominations'].append(category)
        golden_globes['film'].append(film)
        golden_globes['nominee'].append(nominee)
        golden_globes['Golden Globes winner'].append(winner)
        golden_globes['year_award'].append(year_award)
        golden_globes['year_film'].append(year_film)
        nominees_f = nom_cards[n].find_all('li', class_='c-nomination-card__nominee js-nomination-nominee')
        for nf in range(len(nominees_f)):
            film = nominees_f[nf].find('div', class_='u-type-h11').text.strip()
            nominee = nominees_f[nf].find('a', class_='c-nomination-card__name c-nomination-card__link u-color-black-01').text.strip()
            winner = False
            golden_globes['Golden Globes nominations'].append(category)
            golden_globes['film'].append(film)
            golden_globes['nominee'].append(nominee)
            golden_globes['Golden Globes winner'].append(winner)
            golden_globes['year_award'].append(year_award)
            golden_globes['year_film'].append(year_film)
golden_globes

df_golden_extra = pd.DataFrame(golden_globes)

df_golden_extra.tail(50)

#df_golden_extra.to_csv('golden_scrapping.csv', index = False)
