==> notebooks/4_golden_scrap_cleaning.ipynb <==
import pandas as pd
import requests
import re
import cleaning_functions as cf

df_golden_s = pd.read_csv('../data/raw/golden_scrapping.csv')

df_golden_s.head(3)

#seguimos el mismo proceso de limpieza que con el df original de los golden globes

df_golden_s['Golden Globes nominations'] = df_golden_s['Golden Globes nominations'].str.capitalize()

#eliminamos categorías televisivas:
#primero eliminamos las filas con categorías sobre televisión, ya que nuestro estudio se centra en cine/películas
tv_categories = cf.values_iterable_regex(df_golden_s['Golden Globes nominations'],r'\b\w*television\w*\b')
df_golden_s = cf.drop_rows_by_value(df_golden_s,df_golden_s['Golden Globes nominations'],tv_categories)                          

#eliminamos también categrías honoríficas no vinculadas a una película concreta
honorary_awards =['Cecil b. demille award','Henrietta award (world film favorites)','Special achievement award','Carol burnett award']
df_golden_s = cf.drop_rows_by_value(df_golden_s,df_golden_s['Golden Globes nominations'],honorary_awards)

df_golden_s.isnull().sum()

#igual que en el primer DF, los valores nulos de películas se pueden completar con los valores en la misma fila en la columna nominee
nan_rows_golden_s = df_golden_s[df_golden_s.film.isnull()]
nan_rows_golden_s.head(2)

df_golden_s.film.fillna(df_golden_s.nominee, inplace=True)

df_golden_s.isnull().sum()

#unificamos valores únicos en la columna de 'Golden Globes nominations' para igualarlos al df original

df_golden_s['Golden Globes nominations'].unique()

cat_changes = {
    'Best performance by an actress in a motion picture - drama': 'Best actress - drama',
    'Best performance by an actress in a motion picture - musical or comedy': 'Best actress - musical or comedy',
    'Best performance by an actress in a supporting role in any motion picture': 'Best supporting actress - motion picture',
    'Best performance by an actor in a motion picture - drama': 'Best actor - drama',
    'Best performance by an actor in a motion picture - musical or comedy': 'Best actor - musical or comedy',
    'Best performance by an actor in a supporting role in any motion picture': 'Best supporting actor - motion picture',
    'Best motion picture – non-english language': 'Best motion picture - foreign language'
}

df_golden_s = cf.clean_values_column (df_golden_s, cat_changes)
df_golden_s['Golden Globes nominations'].unique()

df_golden_s.head(15)

#df_golden_s.to_csv('golden_scrap_cleaned.csv', index = False)
