==> notebooks/5_golden_unified.ipynb <==
import pandas as pd
import re
import cleaning_functions as cf

golden_scrap =pd.read_csv('../data/Clean/golden/golden_scrap_cleaned.csv')

golden_df = pd.read_csv('../data/Clean/golden/golden_cleaned.csv')

df_golden_final = pd.concat([golden_df,golden_scrap], ignore_index = True)

df_golden_final.tail(15)

df_golden_final.nunique()

#los títulos que empiezan por 'The' o 'A ' lo tienen al final después de una coma. Este formato es distinto que el del df de los oscars e IMDb

updated_film = []
film_list = df_golden_final.film.to_list()

for x in df_golden_final.film:
    if re.search(r'\b\w*, The$', x, flags=re.IGNORECASE):
        y = 'The '+(re.sub(r', The$','',x, flags=re.IGNORECASE))
        updated_film.append(y)
    elif re.search(r'\b\w*, An$', x, flags=re.IGNORECASE):
        y = 'An '+(re.sub(r', An$','',x, flags=re.IGNORECASE))
        updated_film.append(y)
    elif re.search(r'\b\w*, A$', x, flags=re.IGNORECASE):
        y = 'A '+(re.sub(r', A$','',x, flags=re.IGNORECASE))
        updated_film.append(y)
    elif re.search(r'Star Is Born, A', x, flags=re.IGNORECASE):
        y = re.sub(r'Star Is Born, A.+$','A Star Is Born',x, flags=re.IGNORECASE)
        updated_film.append(y)
    else:
        updated_film.append(x)

values_replace_dictio = dict(zip(film_list,updated_film))

df_golden_final.replace(values_replace_dictio, inplace = True)

df_golden_final.tail(6)

#df_golden_final.to_csv('2_golden_final.csv', index = False)
