==> notebooks/6_prizes_unified.ipynb <==
import pandas as pd
import re
import cleaning_functions as cf

oscars = pd.read_csv('../data/Clean/1_oscars_cleaned.csv')
oscars.head(2)

oscars.replace('Star Wars', 'Star Wars: Episode IV - A New Hope', inplace = True)

golden = pd.read_csv('../data/Clean/2_golden_final.csv')
golden.head(2)

#renombramos algunas columnas para mantener los datos claros después de unirlas:
golden.rename(columns={"nominee": "G.G. name", 'Golden Globes nominations':'G.G. nominations', 'Golden Globes winner':'G.G. winner'}, inplace=True)

oscars.rename(columns={"name": "Oscars name"}, inplace=True)

golden.rename(columns={"year_award": "year_ceremony"}, inplace=True)

golden.head(1)

oscars.head(1)

oscars.reset_index(drop = True, inplace = True)

#cambiamos el formato de los df con un pivot table para poder tener las categorías como nombres de columna expandidos para cada película (unmelt)

oscars_films = oscars.pivot_table(index= ['film','year_film','year_ceremony'], columns = 'Oscars nominations', values= 'Oscars winner')
oscars_films = oscars_films.reset_index()
oscars_films.columns.name = None
oscars_films.head(3)

oscars_columns_sorter = ['film', 'year_film', 'year_ceremony', 'Actor in a leading role',
       'Actor in a supporting role', 'Actress in a leading role',
       'Actress in a supporting role', 'Best picture', 'Director', 'Adapted screenplay',
       'Animated feature film', 'Cinematography',
       'Costume design', 'Documentary feature',
       'Documentary feature film', 'Documentary short film', 'Film editing',
       'International feature film', 'Makeup and hairstyling',
       'Original score', 'Original screenplay', 'Original song',
       'Production design', 'Short film animated', 'Short film live action',
       'Sound effects', 'Visual effects']

oscars_films = oscars_films.reindex(columns=oscars_columns_sorter)

oscars_films.head(1)

golden_films = golden.pivot_table(index= ['film','year_film','year_ceremony'], columns = 'G.G. nominations', values= 'G.G. winner')
golden_films = golden_films.reset_index()
golden_films.columns.name = None
golden_films.head(2)

golden_films_sorter = ['film', 'year_film','year_ceremony','Best actor - drama','Best actor - musical or comedy', 'Best actress - drama', 'Best actress - musical or comedy', 'Best director - motion picture', 'Best motion picture - drama', 'Best supporting actor - motion picture', 'Best supporting actress - motion picture', 'Best motion picture - musical or comedy',
       'Best motion picture - animated',
       'Best motion picture - foreign language',
       'Best original score - motion picture',
       'Best original song - motion picture',
       'Best screenplay - motion picture', 'Documentary']

golden_films = golden_films.reindex(columns=golden_films_sorter)

#creamos tres columnas nuevas por cada df: 
    #número de nominaciones (sumar columnas con valor distinto de NaN)
    #número de premios (sumar fila)
    #y si han recibido el premio a la mejor película - posteriormente cambiaremos NaN por 0

oscars_total_nominations = []
total_oscars_won = []
main_categories_nominations = []
main_categories_won= []
oscars_best_picture = []
for n in range(len(oscars_films)):
    oscars_total_nominations.append(int(oscars_films.iloc[n,3:].notnull().sum()))
    total_oscars_won.append(int(oscars_films.iloc[n,3:].sum()))
    main_categories_nominations.append(int(oscars_films.iloc[n,3:9].notnull().sum()))
    main_categories_won.append(int(oscars_films.iloc[n,3:9].sum()))
    if oscars_films.loc[n,'Best picture'] == True:
        oscars_best_picture.append(1)
    else:
        oscars_best_picture.append(0)

oscars_films['Total oscars nominations'] = oscars_total_nominations
oscars_films['Total oscars won'] = total_oscars_won
oscars_films['Oscars main categories nominations'] = main_categories_nominations
oscars_films['Oscars main categories won'] = main_categories_won
oscars_films['Oscars best picture'] = oscars_best_picture
oscars_films.head(3)

golden_total_nominations = []
total_golden_won = []
main_categories_nominations_g = []
main_categories_won_g = []
golden_best_picture = []
for n in range(len(golden_films)):
    golden_total_nominations.append(int(golden_films.iloc[n,3:].notnull().sum()))
    total_golden_won.append(int(golden_films.iloc[n,3:].sum()))
    main_categories_nominations_g.append(int(golden_films.iloc[n,3:12].notnull().sum()))
    main_categories_won_g.append(int(golden_films.iloc[n,3:12].sum()))
    if golden_films.loc[n,'Best motion picture - drama'] == True:
        golden_best_picture.append(1)
    elif golden_films.loc[n,'Best motion picture - musical or comedy'] == True:
        golden_best_picture.append(1)
    else:
        golden_best_picture.append(0)

golden_films['G.G. nominations'] = golden_total_nominations
golden_films['G.G. won'] = total_golden_won
golden_films['G.G. main categories nominations'] = main_categories_nominations_g
golden_films['G.G. main categories won'] = main_categories_won_g
golden_films['G.G. best motion picture'] = golden_best_picture
golden_films.head(2)

#creamos un subset de cada df solo con las columnas que nos interesan para juntarlas en un último dataframe

subset_oscars = oscars_films[['film','year_film','Total oscars nominations','Total oscars won','Oscars main categories nominations','Oscars main categories won','Oscars best picture']]
subset_golden = golden_films[['film','year_film', 'G.G. nominations', 'G.G. won','G.G. main categories nominations','G.G. main categories won','G.G. best motion picture']]

award_winning_movies = pd.merge(subset_oscars,subset_golden, how='outer', on = ['film','year_film'])
award_winning_movies = award_winning_movies.sort_values(by='year_film')
award_winning_movies.reset_index(drop = True, inplace = True)
award_winning_movies.head(2)

#award_winning_movies.to_csv('3_award_winning_movies.csv', index = False)
