==> notebooks/7_tmdb_data.ipynb <==
import requests
from bs4 import BeautifulSoup as bs
import re
import cleaning_functions as cf
import pandas as pd
from tqdm.notebook import tqdm
import passwords

#Ahora extraeremos un nuevo df con información sobre el rating de las películas que ya tenemos de la API de TMDb.
#Para eso necesitamos una lista con las películas a buscar (la sacamos de award_winning_movies) cambiando espacios y caracteres no alfanuméricos por una clave concreta

award_winning_movies = pd.read_csv('../data/Clean/3_award_winning_movies.csv')

movies = pd.DataFrame(award_winning_movies.film)

replace_movies_dictio = {
    "'":'%27',
    " ":'%20',
    ",":'%2C',
    "!":'%09',
    "(":'%28',
    ")":'%29',
    ":":'%3A',
    "?":'%3F',
    "$":'%24',
    "[":'%5B',
    "]":'%5D'
}

movies = cf.clean_strings_column (movies.film, replace_movies_dictio)

films_list = movies.to_list()

len(films_list)

token = passwords.token

url = "https://api.themoviedb.org/3/search/movie?query=A%20Star%20Is%20Born&include_adult=false&language=en-US&page=1"

headers = {
    "accept": "application/json",
    "Authorization": f"Bearer {token}"
}

response = requests.get(url, headers=headers)
film = response.json()

score = film['results'][0]['vote_average']

url = "https://api.themoviedb.org/3/search/movie?query=Far%20from%20the%20Madding%20Crowd&include_adult=false&language=en-US&page=1"

headers = {
    "accept": "application/json",
    "Authorization": f"Bearer {token}"
}

response = requests.get(url, headers=headers)
film = response.json()

film['results'][0]['vote_average']

tmdb_score = []
headers = {
    "accept": "application/json",
    "Authorization": f"Bearer {token}"
}

for film in tqdm(films_list):
    url = f'https://api.themoviedb.org/3/search/movie?query={film}&include_adult=false&language=en-US&page=1'
    response = requests.get(url, headers=headers)
    film_info = response.json()
    try:
        tmdb_score.append(film_info['results'][0]['vote_average'])
    except:
        tmdb_score.append('Not available')
    
len(tmdb_score)

award_winning_movies['TMDb Score'] = tmdb_score

award_winning_movies

award_winning_movies.fillna(0,inplace=True)
award_winning_movies

#award_winning_movies.to_csv('4_final_movie_data.csv', index = False)
